Remove old hexagon layer tests
