+++ test/src/core-layers/hexagon-layer.spec.js
@@ -22,11 +22,7 @@ import test from 'tape-catch';
 import {makeSpy} from 'probe.gl/test';
 
 import * as data from 'deck.gl/test/data';
-import {
-  testInitializeLayer,
-  testLayerUpdates,
-  testSubLayerUpdateTriggers
-} from 'deck.gl-test-utils';
+import {testLayer, testInitializeLayer} from 'deck.gl-test-utils';
 
 import {HexagonLayer, HexagonCellLayer} from 'deck.gl';
 
@@ -34,397 +30,437 @@ const getColorValue = points => points.length;
 const getElevationValue = points => points.length;
 const getPosition = d => d.COORDINATES;
 
-const TEST_CASES = {
-  // props to initialize layer with
-  INITIAL_PROPS: {
-    data: data.points,
-    radius: 400,
-    getPosition
-  },
-  // list of update props to call and asserts on the resulting layer
-  UPDATES: [
-    {
-      updateProps: {
-        radius: 800
+// props to initialize layer with
+// update props
+// asserts on the resulting layer
+test('HexagonLayer#updateLayer', t => {
+  testLayer({
+    Layer: HexagonLayer,
+    userData: t,
+    testCases: [
+      {
+        title: 'Initialize',
+        props: {
+          data: data.points,
+          radius: 400,
+          getPosition
+        }
       },
-      assert: (layer, oldState, t) => {
-        t.ok(oldState.hexagons !== layer.state.hexagons, 'should update layer data');
+      {
+        title: 'Update radius',
+        props: {
+          radius: 800
+        },
+        assert({layer, oldState, userData}) {
+          t.ok(oldState.hexagons !== layer.state.hexagons, 'should update layer data');
 
-        t.ok(
-          oldState.sortedColorBins !== layer.state.sortedColorBins,
-          'should update sortedColorBins'
-        );
+          t.ok(
+            oldState.sortedColorBins !== layer.state.sortedColorBins,
+            'should update sortedColorBins'
+          );
 
-        t.ok(
-          oldState.colorValueDomain !== layer.state.colorValueDomain,
-          'should update valueDomain'
-        );
+          t.ok(
+            oldState.colorValueDomain !== layer.state.colorValueDomain,
+            'should update valueDomain'
+          );
 
-        t.ok(
-          oldState.colorScaleFunc !== layer.state.colorScaleFunc,
-          'should update colorScaleFunc'
-        );
+          t.ok(
+            oldState.colorScaleFunc !== layer.state.colorScaleFunc,
+            'should update colorScaleFunc'
+          );
 
-        t.ok(
-          oldState.sortedElevationBins !== layer.state.sortedElevationBins,
-          'should update sortedElevationBins'
-        );
+          t.ok(
+            oldState.sortedElevationBins !== layer.state.sortedElevationBins,
+            'should update sortedElevationBins'
+          );
 
-        t.ok(
-          oldState.elevationValueDomain !== layer.state.elevationValueDomain,
-          'should update elevationValueDomain'
-        );
+          t.ok(
+            oldState.elevationValueDomain !== layer.state.elevationValueDomain,
+            'should update elevationValueDomain'
+          );
 
-        t.ok(
-          oldState.elevationScaleFunc !== layer.state.elevationScaleFunc,
-          'should update elevationScaleFunc'
-        );
-      }
-    },
-    {
-      updateProps: {
-        getColorValue
+          t.ok(
+            oldState.elevationScaleFunc !== layer.state.elevationScaleFunc,
+            'should update elevationScaleFunc'
+          );
+        }
       },
-      assert: (layer, oldState, t) => {
-        t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
+      {
+        title: 'Update getColorValue accessor',
+        props: {
+          getColorValue
+        },
+        assert({layer, oldState, userData}) {
+          t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
 
-        t.ok(
-          oldState.sortedColorBins !== layer.state.sortedColorBins,
-          'should not update sortedColorBins'
-        );
+          t.ok(
+            oldState.sortedColorBins !== layer.state.sortedColorBins,
+            'should not update sortedColorBins'
+          );
 
-        t.ok(
-          oldState.sortedElevationBins === layer.state.sortedElevationBins,
-          'should update sortedElevationBins'
-        );
+          t.ok(
+            oldState.sortedElevationBins === layer.state.sortedElevationBins,
+            'should update sortedElevationBins'
+          );
 
-        t.ok(
-          oldState.colorValueDomain !== layer.state.colorValueDomain,
-          'should re calculate colorValueDomain'
-        );
+          t.ok(
+            oldState.colorValueDomain !== layer.state.colorValueDomain,
+            'should re calculate colorValueDomain'
+          );
 
-        t.ok(
-          oldState.elevationValueDomain === layer.state.elevationValueDomain,
-          'should not update elevationValueDomain'
-        );
+          t.ok(
+            oldState.elevationValueDomain === layer.state.elevationValueDomain,
+            'should not update elevationValueDomain'
+          );
 
-        t.ok(
-          oldState.colorScaleFunc !== layer.state.colorScaleFunc,
-          'should update colorScaleFunc'
-        );
+          t.ok(
+            oldState.colorScaleFunc !== layer.state.colorScaleFunc,
+            'should update colorScaleFunc'
+          );
 
-        t.ok(
-          oldState.elevationScaleFunc === layer.state.elevationScaleFunc,
-          'should not update colorScaleFunc'
-        );
-      }
-    },
-    {
-      updateProps: {
-        upperPercentile: 90
+          t.ok(
+            oldState.elevationScaleFunc === layer.state.elevationScaleFunc,
+            'should not update colorScaleFunc'
+          );
+        }
       },
-      assert: (layer, oldState, t) => {
-        t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
+      {
+        title: 'Update getColorValue accessor',
+        props: {
+          upperPercentile: 90
+        },
+        assert({layer, oldState, userData}) {
+          t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
 
-        t.ok(
-          oldState.sortedColorBins === layer.state.sortedColorBins,
-          'should not update sortedColorBins'
-        );
+          t.ok(
+            oldState.sortedColorBins === layer.state.sortedColorBins,
+            'should not update sortedColorBins'
+          );
 
-        t.ok(
-          oldState.colorValueDomain !== layer.state.colorValueDomain,
-          'should re calculate colorValueDomain'
-        );
+          t.ok(
+            oldState.colorValueDomain !== layer.state.colorValueDomain,
+            'should re calculate colorValueDomain'
+          );
 
-        t.ok(
-          oldState.colorScaleFunc !== layer.state.colorScaleFunc,
-          'should update colorScaleFunc'
-        );
+          t.ok(
+            oldState.colorScaleFunc !== layer.state.colorScaleFunc,
+            'should update colorScaleFunc'
+          );
 
-        t.ok(
-          oldState.sortedElevationBins === layer.state.sortedElevationBins,
-          'should not update sortedElevationBins'
-        );
+          t.ok(
+            oldState.sortedElevationBins === layer.state.sortedElevationBins,
+            'should not update sortedElevationBins'
+          );
 
-        t.ok(
-          oldState.elevationValueDomain === layer.state.elevationValueDomain,
-          'should not update elevationValueDomain'
-        );
+          t.ok(
+            oldState.elevationValueDomain === layer.state.elevationValueDomain,
+            'should not update elevationValueDomain'
+          );
 
-        t.ok(
-          oldState.elevationScaleFunc === layer.state.elevationScaleFunc,
-          'should not update elevationScaleFunc'
-        );
-      }
-    },
-    {
-      updateProps: {
-        colorDomain: [0, 10]
+          t.ok(
+            oldState.elevationScaleFunc === layer.state.elevationScaleFunc,
+            'should not update elevationScaleFunc'
+          );
+        }
       },
-      assert: (layer, oldState, t) => {
-        t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
+      {
+        title: 'Update colorDomain',
+        props: {
+          colorDomain: [0, 10]
+        },
+        assert({layer, oldState, userData}) {
+          t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
 
-        t.ok(
-          oldState.sortedColorBins === layer.state.sortedColorBins,
-          'should not update sortedColorBins'
-        );
+          t.ok(
+            oldState.sortedColorBins === layer.state.sortedColorBins,
+            'should not update sortedColorBins'
+          );
 
-        t.ok(
-          oldState.colorValueDomain === layer.state.colorValueDomain,
-          'should not re calculate colorValueDomain'
-        );
+          t.ok(
+            oldState.colorValueDomain === layer.state.colorValueDomain,
+            'should not re calculate colorValueDomain'
+          );
 
-        t.ok(
-          oldState.colorScaleFunc !== layer.state.colorScaleFunc,
-          'should update colorScaleFunc'
-        );
+          t.ok(
+            oldState.colorScaleFunc !== layer.state.colorScaleFunc,
+            'should update colorScaleFunc'
+          );
 
-        t.ok(
-          oldState.sortedElevationBins === layer.state.sortedElevationBins,
-          'should not update sortedElevationBins'
-        );
+          t.ok(
+            oldState.sortedElevationBins === layer.state.sortedElevationBins,
+            'should not update sortedElevationBins'
+          );
 
-        t.ok(
-          oldState.elevationValueDomain === layer.state.elevationValueDomain,
-          'should not update elevationValueDomain'
-        );
+          t.ok(
+            oldState.elevationValueDomain === layer.state.elevationValueDomain,
+            'should not update elevationValueDomain'
+          );
 
-        t.ok(
-          oldState.elevationScaleFunc === layer.state.elevationScaleFunc,
-          'should not update elevationScaleFunc'
-        );
-      }
-    },
-    {
-      updateProps: {
-        getElevationValue
+          t.ok(
+            oldState.elevationScaleFunc === layer.state.elevationScaleFunc,
+            'should not update elevationScaleFunc'
+          );
+        }
       },
-      assert: (layer, oldState, t) => {
-        t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
+      {
+        title: 'Update getElevationValue accessor',
+        props: {
+          getElevationValue
+        },
+        assert({layer, oldState, userData}) {
+          t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
 
-        t.ok(
-          oldState.sortedElevationBins !== layer.state.sortedElevationBins,
-          'should update sortedElevationBins'
-        );
+          t.ok(
+            oldState.sortedElevationBins !== layer.state.sortedElevationBins,
+            'should update sortedElevationBins'
+          );
 
-        t.ok(
-          oldState.elevationValueDomain !== layer.state.elevationValueDomain,
-          'should re calculate elevationValueDomain'
-        );
+          t.ok(
+            oldState.elevationValueDomain !== layer.state.elevationValueDomain,
+            'should re calculate elevationValueDomain'
+          );
 
-        t.ok(
-          oldState.elevationScaleFunc !== layer.state.elevationScaleFunc,
-          'should update elevationScaleFunc'
-        );
+          t.ok(
+            oldState.elevationScaleFunc !== layer.state.elevationScaleFunc,
+            'should update elevationScaleFunc'
+          );
 
-        t.ok(
-          oldState.sortedColorBins === layer.state.sortedColorBins,
-          'should not update sortedColorBins'
-        );
+          t.ok(
+            oldState.sortedColorBins === layer.state.sortedColorBins,
+            'should not update sortedColorBins'
+          );
 
-        t.ok(
-          oldState.colorValueDomain === layer.state.colorValueDomain,
-          'should not re calculate colorValueDomain'
-        );
+          t.ok(
+            oldState.colorValueDomain === layer.state.colorValueDomain,
+            'should not re calculate colorValueDomain'
+          );
 
-        t.ok(
-          oldState.colorScaleFunc === layer.state.colorScaleFunc,
-          'should not update colorScaleFunc'
-        );
-      }
-    },
-    {
-      updateProps: {
-        elevationLowerPercentile: 1
+          t.ok(
+            oldState.colorScaleFunc === layer.state.colorScaleFunc,
+            'should not update colorScaleFunc'
+          );
+        }
       },
-      assert: (layer, oldState, t) => {
-        t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
+      {
+        title: 'Update elevation lower percentile',
+        props: {
+          elevationLowerPercentile: 1
+        },
+        assert({layer, oldState, userData}) {
+          t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
 
-        t.ok(
-          oldState.sortedElevationBins === layer.state.sortedElevationBins,
-          'should not update sortedElevationBins'
-        );
+          t.ok(
+            oldState.sortedElevationBins === layer.state.sortedElevationBins,
+            'should not update sortedElevationBins'
+          );
 
-        t.ok(
-          oldState.elevationValueDomain !== layer.state.elevationValueDomain,
-          'should re calculate elevationValueDomain'
-        );
+          t.ok(
+            oldState.elevationValueDomain !== layer.state.elevationValueDomain,
+            'should re calculate elevationValueDomain'
+          );
 
-        t.ok(
-          oldState.elevationScaleFunc !== layer.state.elevationScaleFunc,
-          'should update elevationScaleFunc'
-        );
+          t.ok(
+            oldState.elevationScaleFunc !== layer.state.elevationScaleFunc,
+            'should update elevationScaleFunc'
+          );
 
-        t.ok(
-          oldState.sortedColorBins === layer.state.sortedColorBins,
-          'should not update sortedColorBins'
-        );
+          t.ok(
+            oldState.sortedColorBins === layer.state.sortedColorBins,
+            'should not update sortedColorBins'
+          );
 
-        t.ok(
-          oldState.colorValueDomain === layer.state.colorValueDomain,
-          'should not re calculate colorValueDomain'
-        );
+          t.ok(
+            oldState.colorValueDomain === layer.state.colorValueDomain,
+            'should not re calculate colorValueDomain'
+          );
 
-        t.ok(
-          oldState.colorScaleFunc === layer.state.colorScaleFunc,
-          'should not update colorScaleFunc'
-        );
-      }
-    },
-    {
-      updateProps: {
-        elevationRange: [1, 10]
+          t.ok(
+            oldState.colorScaleFunc === layer.state.colorScaleFunc,
+            'should not update colorScaleFunc'
+          );
+        }
       },
-      assert: (layer, oldState, t) => {
-        t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
+      {
+        title: 'Update elevationRange accessor',
+        props: {
+          elevationRange: [1, 10]
+        },
+        assert({layer, oldState, userData}) {
+          t.ok(oldState.hexagons === layer.state.hexagons, 'should not update layer data');
 
-        t.ok(
-          oldState.sortedElevationBins === layer.state.sortedElevationBins,
-          'should not update sortedElevationBins'
-        );
+          t.ok(
+            oldState.sortedElevationBins === layer.state.sortedElevationBins,
+            'should not update sortedElevationBins'
+          );
 
-        t.ok(
-          oldState.elevationValueDomain === layer.state.elevationValueDomain,
-          'should not re calculate elevationValueDomain'
-        );
+          t.ok(
+            oldState.elevationValueDomain === layer.state.elevationValueDomain,
+            'should not re calculate elevationValueDomain'
+          );
 
-        t.ok(
-          oldState.elevationScaleFunc !== layer.state.elevationScaleFunc,
-          'should update elevationScaleFunc'
-        );
+          t.ok(
+            oldState.elevationScaleFunc !== layer.state.elevationScaleFunc,
+            'should update elevationScaleFunc'
+          );
 
-        t.ok(
-          oldState.sortedColorBins === layer.state.sortedColorBins,
-          'should not update sortedColorBins'
-        );
+          t.ok(
+            oldState.sortedColorBins === layer.state.sortedColorBins,
+            'should not update sortedColorBins'
+          );
 
-        t.ok(
-          oldState.colorValueDomain === layer.state.colorValueDomain,
-          'should not re calculate colorValueDomain'
-        );
+          t.ok(
+            oldState.colorValueDomain === layer.state.colorValueDomain,
+            'should not re calculate colorValueDomain'
+          );
 
-        t.ok(
-          oldState.colorScaleFunc === layer.state.colorScaleFunc,
-          'should not update colorScaleFunc'
-        );
+          t.ok(
+            oldState.colorScaleFunc === layer.state.colorScaleFunc,
+            'should not update colorScaleFunc'
+          );
+        }
       }
-    }
-  ]
-};
+    ]
+  });
 
-const SUBLAYER_TEST_CASES = {
-  // props to initialize layer with
-  INITIAL_PROPS: {
-    data: data.points,
-    radius: 400,
-    getPosition
-  },
-  // list of update props to call and asserts on the resulting layer
-  UPDATES: [
-    {
-      updateProps: {
-        radius: 800
+  t.end();
+});
+
+test('HexagonLayer#updateTriggers', t => {
+  const functionsToSpy = ['_onGetSublayerColor', '_onGetSublayerElevation'];
+
+  testLayer({
+    Layer: HexagonLayer,
+    spies: functionsToSpy,
+    userData: t,
+    testCases: [
+      {
+        // props to initialize layer with
+        props: {
+          data: data.points,
+          radius: 400,
+          getPosition
+        }
       },
-      assert: (subLayer, spies, t) => {
-        t.ok(spies._onGetSublayerColor.called, 'update radius should call _onGetSublayerColor');
-        t.ok(
-          spies._onGetSublayerElevation.called,
-          'update radius should call _onGetSublayerElevation'
-        );
-      }
-    },
-    {
-      updateProps: {
-        opacity: 0.1
+      {
+        title: 'Update radius prop',
+        props: {
+          radius: 800
+        },
+        spies: functionsToSpy,
+        assert({subLayer, spies, userData}) {
+          t.ok(spies._onGetSublayerColor.called, 'update radius should call _onGetSublayerColor');
+          t.ok(
+            spies._onGetSublayerElevation.called,
+            'update radius should call _onGetSublayerElevation'
+          );
+        }
       },
-      assert: (subLayer, spies, t) => {
-        t.ok(
-          !spies._onGetSublayerColor.called,
-          'update opacity should not call _onGetSublayerColor'
-        );
-        t.ok(
-          !spies._onGetSublayerElevation.called,
-          'update opacity  should not call _onGetSublayerElevation'
-        );
-      }
-    },
-    {
-      updateProps: {
-        getColorValue
+      {
+        title: 'Update opacity prop',
+        props: {
+          opacity: 0.1
+        },
+        spies: functionsToSpy,
+        assert({subLayer, spies, userData}) {
+          t.ok(
+            !spies._onGetSublayerColor.called,
+            'update opacity should not call _onGetSublayerColor'
+          );
+          t.ok(
+            !spies._onGetSublayerElevation.called,
+            'update opacity  should not call _onGetSublayerElevation'
+          );
+        }
       },
-      assert: (subLayer, spies, t) => {
-        t.ok(
-          spies._onGetSublayerColor.called,
-          'update getColorValue should call _onGetSublayerColor'
-        );
-        t.ok(
-          !spies._onGetSublayerElevation.called,
-          'update getColorValue should not call _onGetSublayerElevation'
-        );
-      }
-    },
-    {
-      updateProps: {
-        upperPercentile: 90
+      {
+        title: 'Update getColorValue prop',
+        props: {
+          getColorValue
+        },
+        spies: functionsToSpy,
+        assert({subLayer, spies, userData}) {
+          t.ok(
+            spies._onGetSublayerColor.called,
+            'update getColorValue should call _onGetSublayerColor'
+          );
+          t.ok(
+            !spies._onGetSublayerElevation.called,
+            'update getColorValue should not call _onGetSublayerElevation'
+          );
+        }
       },
-      assert: (subLayer, spies, t) => {
-        t.ok(
-          spies._onGetSublayerColor.called,
-          'update upperPercentile should call _onGetSublayerColor'
-        );
-        t.ok(
-          !spies._onGetSublayerElevation.called,
-          'update upperPercentile should not call _onGetSublayerElevation'
-        );
-      }
-    },
-    {
-      updateProps: {
-        getElevationValue
+      {
+        title: 'Update upperPercentile prop',
+        props: {
+          upperPercentile: 90
+        },
+        spies: functionsToSpy,
+        assert({subLayer, spies, userData}) {
+          t.ok(
+            spies._onGetSublayerColor.called,
+            'update upperPercentile should call _onGetSublayerColor'
+          );
+          t.ok(
+            !spies._onGetSublayerElevation.called,
+            'update upperPercentile should not call _onGetSublayerElevation'
+          );
+        }
       },
-      assert: (subLayer, spies, t) => {
-        t.ok(
-          !spies._onGetSublayerColor.called,
-          'update getElevationValue should not call _onGetSublayerColor'
-        );
-        t.ok(
-          spies._onGetSublayerElevation.called,
-          'update getElevationValue should call _onGetSublayerElevation'
-        );
-      }
-    },
-    {
-      updateProps: {
-        elevationUpperPercentile: 99
+      {
+        title: 'Update getElevationValue prop',
+        props: {
+          getElevationValue
+        },
+        spies: functionsToSpy,
+        assert({subLayer, spies, userData}) {
+          t.ok(
+            !spies._onGetSublayerColor.called,
+            'update getElevationValue should not call _onGetSublayerColor'
+          );
+          t.ok(
+            spies._onGetSublayerElevation.called,
+            'update getElevationValue should call _onGetSublayerElevation'
+          );
+        }
       },
-      assert: (subLayer, spies, t) => {
-        t.ok(
-          !spies._onGetSublayerColor.called,
-          'update elevationUpperPercentile should not call _onGetSublayerColor'
-        );
-        t.ok(
-          spies._onGetSublayerElevation.called,
-          'update elevationUpperPercentile should call _onGetSublayerElevation'
-        );
-      }
-    },
-    {
-      updateProps: {
-        elevationRange: [0, 100]
+      {
+        title: 'Update elevationUpperPercentile prop',
+        props: {
+          elevationUpperPercentile: 99
+        },
+        spies: functionsToSpy,
+        assert({subLayer, spies, userData}) {
+          t.ok(
+            !spies._onGetSublayerColor.called,
+            'update elevationUpperPercentile should not call _onGetSublayerColor'
+          );
+          t.ok(
+            spies._onGetSublayerElevation.called,
+            'update elevationUpperPercentile should call _onGetSublayerElevation'
+          );
+        }
       },
-      assert: (subLayer, spies, t) => {
-        t.ok(
-          !spies._onGetSublayerColor.called,
-          'update elevationRange should not call _onGetSublayerColor'
-        );
-        t.ok(
-          spies._onGetSublayerElevation.called,
-          'update elevationRange should call _onGetSublayerElevation'
-        );
+      {
+        title: 'Update elevationRange prop',
+        props: {
+          elevationRange: [0, 100]
+        },
+        spies: functionsToSpy,
+        assert({subLayer, spies, userData}) {
+          t.ok(
+            !spies._onGetSublayerColor.called,
+            'update elevationRange should not call _onGetSublayerColor'
+          );
+          t.ok(
+            spies._onGetSublayerElevation.called,
+            'update elevationRange should call _onGetSublayerElevation'
+          );
+        }
       }
-    }
-  ]
-};
+    ]
+  });
+  t.end();
+});
 
 test('HexagonLayer#constructor', t => {
   let layer = new HexagonLayer({
@@ -496,23 +532,6 @@ test('HexagonLayer#renderSubLayer', t => {
   );
   HexagonLayer.prototype._onGetSublayerColor.restore();
   HexagonLayer.prototype._onGetSublayerElevation.restore();
-
-  t.end();
-});
-
-test('HexagonLayer#updateLayer', t => {
-  testLayerUpdates(t, {LayerComponent: HexagonLayer, testCases: TEST_CASES});
-  t.end();
-});
-
-test('HexagonLayer#updateTriggers', t => {
-  const FunctionsToSpy = ['_onGetSublayerColor', '_onGetSublayerElevation'];
-
-  testSubLayerUpdateTriggers(t, {
-    FunctionsToSpy,
-    LayerComponent: HexagonLayer,
-    testCases: SUBLAYER_TEST_CASES
-  });
 
   t.end();
 });
